+++ tests/compat/test_numba.py
@@ -1,15 +1,23 @@
 from brtp.compat import is_numba_installed, numba
+from brtp.compat._numba._dummy_numba import Numba
 
 
 def test_is_numba_installed():
     # --- arrange -----------------------------------------
-    is_numba_truly_installed = hasattr(numba, "double")
+    try:
+        import numba as real_numba
+
+        is_numba_truly_installed = True
+    except ImportError:
+        is_numba_truly_installed = False
 
     # --- act ---------------------------------------------
     is_numba_deemed_installed = is_numba_installed()
 
     # --- assert ------------------------------------------
     assert is_numba_deemed_installed == is_numba_truly_installed
+    if not is_numba_deemed_installed:
+        assert not isinstance(numba, Numba), "Dummy numba instance not used when numba is installed."
 
 
 def test_numba_decorator_always_works():
@@ -35,3 +43,52 @@ def test_numba_decorator_always_works():
     assert result_jit == 11, "numba.jit decorator did not work as expected."
     assert result_njit == 12, "numba.njit decorator did not work as expected."
     assert result_njit2 == 13, "numba.njit(...) decorator did not work as expected."
+
+
+def test_dummy_numba_typed_dict():
+    # --- arrange -----------------------------------------
+    @numba.njit
+    def try_typed_dict() -> float:
+        typed_dict = numba.typed.Dict.empty(
+            key_type=numba.types.int64,
+            value_type=numba.types.float64,
+        )
+        typed_dict[1] = 2.71
+        typed_dict[1] = 3.14
+        return typed_dict[1]
+
+    # --- act ---------------------------------------------
+    result = try_typed_dict()
+
+    # --- assert ------------------------------------------
+    assert result == 3.14
+
+
+def test_dummy_numba_typed_list():
+    # --- arrange -----------------------------------------
+    @numba.njit
+    def try_typed_list() -> float:
+        typed_list = numba.typed.List.empty_list(numba.types.float64)
+        typed_list.append(2.71)
+        typed_list.append(3.14)
+        return typed_list[1]
+
+    # --- act ---------------------------------------------
+    result = try_typed_list()
+
+    # --- assert ------------------------------------------
+    assert result == 3.14
+
+
+def test_dummy_numba_typed_list_conversion():
+    # --- arrange -----------------------------------------
+    lst_python = [1.1, 2.2, 3.3]
+
+    # --- act ---------------------------------------------
+    lst_numba = numba.typed.List(lst_python)
+
+    # --- assert ------------------------------------------
+    assert lst_numba[0] == 1.1
+    assert lst_numba[1] == 2.2
+    assert lst_numba[2] == 3.3
+    assert len(lst_numba) == 3